requirements.txt
langchain
langchain-groq
langchain-huggingface
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
pymupdf
unstructured
gradio
langserve
fastapi
uvicorn
sse_starlette
python-dotenv
